+++ hydra-container/src/state.rs
@@ -1,17 +1,17 @@
 use std::{
     collections::{HashMap, HashSet},
-    sync::atomic::{AtomicU64, Ordering},
+    sync::atomic::{AtomicU32, Ordering},
 };
 
 use tokio::sync::mpsc;
 
 use crate::{commands::Command, pty::PtyCommands};
 
-static PTY_ID: AtomicU64 = AtomicU64::new(0);
+static PTY_ID: AtomicU32 = AtomicU32::new(0);
 
 pub struct State {
     commands: mpsc::Sender<Command>,
-    ptys: HashMap<u64, mpsc::Sender<PtyCommands>>,
+    ptys: HashMap<u32, mpsc::Sender<PtyCommands>>,
 }
 
 impl State {
@@ -22,17 +22,17 @@ impl State {
         }
     }
 
-    pub fn get_pty(&mut self, id: u64) -> Option<&mut mpsc::Sender<PtyCommands>> {
+    pub fn get_pty(&mut self, id: u32) -> Option<&mut mpsc::Sender<PtyCommands>> {
         self.ptys.get_mut(&id)
     }
 
-    pub fn register_pty(&mut self, pty: mpsc::Sender<PtyCommands>) -> u64 {
+    pub fn register_pty(&mut self, pty: mpsc::Sender<PtyCommands>) -> u32 {
         let id = PTY_ID.fetch_add(1, Ordering::Relaxed);
         self.ptys.insert(id, pty);
         id
     }
 
-    pub fn remove_pty(&mut self, id: u64) {
+    pub fn remove_pty(&mut self, id: u32) {
         self.ptys.remove(&id);
     }
 }
